more reorg and namespace move
