+++ test/Microsoft.ML.Tests/ScenariosWithDirectInstantiation/SentimentPredictionTests.cs
@@ -46,7 +46,7 @@ namespace Microsoft.ML.Scenarios
                     },
                     OutputTokens = true,
                     KeepPunctuations = false,
-                    StopWordsRemover = new Runtime.TextAnalytics.PredefinedStopWordsRemoverFactory(),
+                    StopWordsRemover = new PredefinedStopWordsRemoverFactory(),
                     VectorNormalizer = TextFeaturizingEstimator.TextNormKind.L2,
                     CharFeatureExtractor = new NgramExtractorTransform.NgramExtractorArguments() { NgramLength = 3, AllLengths = false },
                     WordFeatureExtractor = new NgramExtractorTransform.NgramExtractorArguments() { NgramLength = 2, AllLengths = true },
@@ -109,7 +109,7 @@ namespace Microsoft.ML.Scenarios
                     },
                     OutputTokens = true,
                     KeepPunctuations= false,
-                    StopWordsRemover = new Runtime.TextAnalytics.PredefinedStopWordsRemoverFactory(),
+                    StopWordsRemover = new PredefinedStopWordsRemoverFactory(),
                     VectorNormalizer = TextFeaturizingEstimator.TextNormKind.None,
                     CharFeatureExtractor = null,
                     WordFeatureExtractor = null,
